feat(map): Add draggable Menu with list of books
Added draggable menu that display the title, author,image, rating
and tag of the book.
For now, the tag and image are placeholder as they are not well
defined inside the project.
The draggable Menu cannot disappear of the map screen.
Used constant for the hard coding in the draggable menu,customized
 info window and map screen.
Privatized CustomInfoWindow.