fix(loginTest): address pre-login notification issue
Attempted to resolve the problem of the notification demand that appears before the login process. Wich made the loginTest fail.
Tried implementing an onClick handler for the notification permission in the loginTest directly, but the notification has no composable, text, or test tag, making it unclickable via testing tools.
Further investigation is required, and a complete solution is planned for next week.